requirement.txt
colorama==0.4.6
GitPython==3.1.29
